feat(frontend): split up sidebars in sub-components, +more work on profile page
